Merge both versions
The costume physic engine has been added
(notice: some functions are disable for now until the game is set up and has a strong base so they dont have problems with new incoming functions)
Menu system: the menu system works more or less like an array, from 0 to X, and the last scene is suppose to teake you to the first one or quit the game. right now all the screens are connected, except the test scene that is not accessable through the game but only through the engine.